fix: align entity types with the provided SQL schema
id was mapped as Long (expects bigint) but the script defines int
columns. Rental.surface/price were Double (expects float) but the
columns are decimal. Hibernate's schema validation rejected both on
startup. Changed id to Integer and surface/price to BigDecimal.

+++ src/main/java/com/chatop/api/entity/Rental.java
@@ -15,6 +15,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "RENTALS")
 @Getter
@@ -26,13 +28,13 @@ public class Rental extends Auditable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
 
-    private Double surface;
+    private BigDecimal surface;
 
-    private Double price;
+    private BigDecimal price;
 
     private String picture;
